refactor: 신입회원 모집 admin, user 기능
- 신입회원 모집 관련 페이지를 admin, user에 따라 구분
- 모집 설정 변경 기능 추가
- 오탈자 수정
